Удаляет ненужный код
